+++ src/generator/c.rs
@@ -150,7 +150,8 @@ fn generate_expression(expr: Expression) -> String {
         Expression::ArrayAccess(name, expr) => generate_array_access(name, *expr),
         Expression::BinOp(left, op, right) => generate_bin_op(*left, op, *right),
         Expression::StructInitialization(_, fields) => generate_struct_initialization(fields),
-        Expression::FieldAccess(expr, field) => generate_field_access(*expr, field),
+        Expression::FieldAccess(expr, field) => generate_field_access(*expr, *field),
+        Expression::Selff => todo!(),
     }
 }
 
@@ -244,7 +245,8 @@ fn generate_function_call(func: String, args: Vec<Expression>) -> String {
             Expression::Array(_) => todo!(),
             Expression::BinOp(left, op, right) => generate_bin_op(*left, op, *right),
             Expression::StructInitialization(_, fields) => generate_struct_initialization(fields),
-            Expression::FieldAccess(expr, field) => generate_field_access(*expr, field),
+            Expression::FieldAccess(expr, field) => generate_field_access(*expr, *field),
+            Expression::Selff => todo!(),
         })
         .collect::<Vec<String>>()
         .join(",");
@@ -298,8 +300,12 @@ fn generate_struct_initialization(fields: HashMap<String, Box<Expression>>) -> S
     buf
 }
 
-fn generate_field_access(expr: Expression, field: String) -> String {
-    format!("{}.{}", generate_expression(expr), field)
+fn generate_field_access(expr: Expression, field: Expression) -> String {
+    format!(
+        "{}.{}",
+        generate_expression(expr),
+        generate_expression(field)
+    )
 }
 
 fn generate_assign(name: Expression, expr: Expression) -> String {